+++ components/data-display/circle-welcome.tsx
@@ -1,42 +1,19 @@
-import { zodResolver } from "@hookform/resolvers/zod"
-import {
-  CameraIcon,
-  ChevronLeftIcon,
-  ChevronRightIcon,
-  CircleHelpIcon,
-  SquarePenIcon,
-} from "@yamada-ui/lucide"
+import { SquarePenIcon } from "@yamada-ui/lucide"
 import type { FC } from "@yamada-ui/react"
 import {
   Button,
-  ButtonGroup,
-  Card,
   CardBody,
   CardHeader,
-  Center,
   Grid,
   Heading,
-  HStack,
-  IconButton,
   Image,
-  Input,
-  Modal,
-  ModalBody,
-  ModalFooter,
-  ModalHeader,
   Text,
-  Textarea,
-  Tooltip,
   useDisclosure,
-  useNotice,
-  useSafeLayoutEffect,
   VStack,
-  FileButton,
 } from "@yamada-ui/react"
 import { useState } from "react"
-import { useForm, Controller } from "react-hook-form"
+import { WelcomeCardForm } from "../forms/welcome-card-form"
 import { WelcomeCard } from "./welcome-card"
-import { FrontWelcomeCardSchema } from "@/schema/welcome"
 import type { FrontWelcomeCard } from "@/schema/welcome"
 
 interface CircleWelcomeProps {
@@ -45,9 +22,6 @@ interface CircleWelcomeProps {
 
 export const CircleWelcome: FC<CircleWelcomeProps> = ({ isMember }) => {
   const { open, onOpen, onClose } = useDisclosure()
-  const notice = useNotice()
-
-  const [currentCard, setCurrentCard] = useState<number>(0)
   const [cards, setCards] = useState<FrontWelcomeCard[]>([
     {
       frontTitle: "",
@@ -68,91 +42,6 @@ export const CircleWelcome: FC<CircleWelcomeProps> = ({ isMember }) => {
       backDescription: "",
     },
   ])
-  const [draftCards, setDraftCards] = useState<FrontWelcomeCard[]>(cards)
-
-  const { register, handleSubmit, reset, watch, control } =
-    useForm<FrontWelcomeCard>({
-      resolver: zodResolver(FrontWelcomeCardSchema),
-      defaultValues: draftCards[currentCard],
-    })
-
-  const watchCards = watch()
-
-  // onSubmitを簡略化（状態の更新は不要）
-  const onSubmit = async (data: FrontWelcomeCard) => {
-    const updatedCards = await Promise.all(
-      draftCards.map(async (card, index) => {
-        if (index === currentCard) return data
-        const result = await FrontWelcomeCardSchema.parseAsync(card)
-        return {
-          frontTitle: result.frontTitle || "",
-          backTitle: result.backTitle || "",
-          backDescription: result.backDescription || "",
-          frontImage: result.frontImage || null,
-        }
-      }),
-    )
-
-    setCards(updatedCards)
-    notice({
-      title: `ウェルカムカードを更新しました`,
-      status: "success",
-      placement: "bottom",
-    })
-  }
-
-  const [imagePreview, setImagePreview] = useState<string>("")
-
-  const handlePrevCard = () => {
-    if (currentCard > 0) {
-      setCurrentCard((prev) => prev - 1)
-      setDraftCards((prev) => {
-        const newCards = [...prev]
-        newCards[currentCard] = watchCards
-        return newCards
-      })
-    }
-  }
-
-  const handleNextCard = () => {
-    if (currentCard < 2) {
-      setCurrentCard((prev) => prev + 1)
-      setDraftCards((prev) => {
-        const newCards = [...prev]
-        newCards[currentCard] = watchCards
-        return newCards
-      })
-    }
-  }
-
-  // モーダルが開くときと、カード切り替え時のみ実行されるように修正
-  useSafeLayoutEffect(() => {
-    reset(draftCards[currentCard])
-    // カードの画像が存在する場合はプレビューを設定
-    if (typeof draftCards[currentCard].frontImage === "string") {
-      setImagePreview(draftCards[currentCard].frontImage)
-    } else {
-      setImagePreview("")
-    }
-  }, [currentCard]) // cardsとresetを依存配列から削除
-
-  // watchでimagePathを監視
-  const imagePath = watch("frontImage") as unknown as FileList | null
-
-  // 画像選択時にプレビューを更新
-  useSafeLayoutEffect(() => {
-    if (typeof imagePath === "string") return
-    if (imagePath && imagePath.length > 0) {
-      const file = imagePath[0]
-      setImagePreview(URL.createObjectURL(file))
-    } else {
-      setImagePreview("")
-    }
-    // クリーンアップ関数でURLを解放
-    return () => {
-      if (imagePreview) URL.revokeObjectURL(imagePreview)
-    }
-  }, [imagePath])
 
   const renderCardContent = (index: number) => {
     const card = cards[index]
@@ -246,124 +135,12 @@ export const CircleWelcome: FC<CircleWelcomeProps> = ({ isMember }) => {
 
   return (
     <VStack w="full" h="full">
-      <Modal
+      <WelcomeCardForm
         open={open}
         onClose={onClose}
-        size="4xl"
-        scrollBehavior="inside"
-        withCloseButton={false}
-      >
-        <ModalHeader justifyContent="space-between">
-          <Text>カード{currentCard + 1}/3</Text>
-          <HStack>
-            <Tooltip
-              label="カードを3つ全て入力するとウェルカムページで反映されます。"
-              placement="bottom"
-            >
-              <Center>
-                <CircleHelpIcon />
-              </Center>
-            </Tooltip>
-            <ButtonGroup attached>
-              <IconButton
-                icon={<ChevronLeftIcon />}
-                colorScheme="riverBlue"
-                onClick={handlePrevCard}
-              />
-              <IconButton
-                icon={<ChevronRightIcon />}
-                colorScheme="riverBlue"
-                onClick={handleNextCard}
-              />
-            </ButtonGroup>
-          </HStack>
-        </ModalHeader>
-        <ModalBody>
-          <HStack
-            w="full"
-            h="full"
-            alignItems="flex-start"
-            mb="xs"
-            flexDir={{ base: "row", md: "column" }}
-          >
-            <VStack w="full" h="full" gap="md">
-              <Text fontSize="md">表面</Text>
-              <Card minH={{ base: "sm", md: "2xs" }}>
-                <CardHeader>
-                  <Input
-                    placeholder="何をするサークル？"
-                    {...register("frontTitle")}
-                  />
-                </CardHeader>
-                <CardBody>
-                  <Center
-                    w="full"
-                    h="full"
-                    flexGrow={1}
-                    {...(imagePreview
-                      ? {
-                          backgroundImage: imagePreview,
-                          backgroundSize: "cover",
-                          backgroundPosition: "center",
-                        }
-                      : {
-                          backgroundColor: "gray.100",
-                        })}
-                  >
-                    <Controller
-                      name="frontImage"
-                      control={control}
-                      render={({ field: { ref, name, onBlur, onChange } }) => (
-                        <FileButton
-                          {...{ ref, name, onBlur }}
-                          w="16"
-                          h="16"
-                          as={IconButton}
-                          accept="image/*"
-                          onChange={onChange}
-                          icon={<CameraIcon fontSize="5xl" color="gray" />}
-                          fullRounded
-                          variant="outline"
-                        />
-                      )}
-                    />
-                  </Center>
-                </CardBody>
-              </Card>
-            </VStack>
-            <VStack w="full" h="full" gap="md">
-              <Text fontSize="md">裏面</Text>
-              <Card minH={{ base: "sm", md: "2xs" }}>
-                <CardHeader>
-                  <Input
-                    placeholder="サークルのメンバーは？"
-                    {...register("backTitle")}
-                  />
-                </CardHeader>
-                <CardBody>
-                  <Textarea
-                    w="full"
-                    h="full"
-                    flexGrow={1}
-                    placeholder="サークルの詳しい説明"
-                    {...register("backDescription")}
-                  />
-                </CardBody>
-              </Card>
-            </VStack>
-          </HStack>
-        </ModalBody>
-        <ModalFooter>
-          <ButtonGroup attached>
-            <Button colorScheme="riverBlue" onClick={onClose}>
-              キャンセル
-            </Button>
-            <Button colorScheme="riverBlue" onClick={handleSubmit(onSubmit)}>
-              更新
-            </Button>
-          </ButtonGroup>
-        </ModalFooter>
-      </Modal>
+        cards={cards}
+        onUpdateCards={setCards}
+      />
       {isMember && (
         <Button
           startIcon={<SquarePenIcon />}
